move metric types to another package
