+++ app/src/main/java/org/astraea/metrics/jmx/types/CompositeDataMetric.java
@@ -1,4 +1,4 @@
-package org.astraea.metrics.jmx;
+package org.astraea.metrics.jmx.types;
 
 public class CompositeDataMetric extends JmxBrokerMetric {
