+++ app/src/main/java/org/astraea/metrics/jmx/types/DoubleBrokerMetric.java
@@ -1,4 +1,4 @@
-package org.astraea.metrics.jmx;
+package org.astraea.metrics.jmx.types;
 
 public class DoubleBrokerMetric extends JmxBrokerMetric {
   public DoubleBrokerMetric(String jmxObjectName, String attributeName)
