+++ app/src/main/java/org/astraea/metrics/jmx/types/IntegerBrokerMetric.java
@@ -1,4 +1,4 @@
-package org.astraea.metrics.jmx;
+package org.astraea.metrics.jmx.types;
 
 public class IntegerBrokerMetric extends JmxBrokerMetric {
